--- network/src/processor/js/exports/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2020 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0
-
-//!
-//! API(s) that are exported to javascript context
-//!
-
-pub mod graphql_api_v1;
